fixed hidden error

+++ cli/train.py
@@ -417,7 +417,8 @@ def main():
 
             print("pred tokens", predicted_tokens.shape, predicted_tokens)
             print("labels", labels.shape, labels)
-            loss = F.cross_entropy(predicted_tokens, labels)
+            
+            loss = F.cross_entropy(predicted_tokens.to(dtype=torch.long), labels)
             optimizer.zero_grad() 
 
             loss.backward()
